requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
dlib>=19.24.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
